Access Policy: Included Pseudo Statements for 'none' policy
Pseudo Statements for 'none' are included to verify that 'none' is a valid policy for given prefix (returns error if there is nesting issue)
The Pseudo none statements are removed immediately after checking. In this way 'none' policy will be treated and checked the same as 'readonly', 'writeonly', and 'readwrite'.