distribute: Make server work with multiple remote disks
This change initializes rpc servers associated with disks that are
local. It makes object layer initialization on demand, namely on the
first request to the object layer.

N B In a multi-node setup, format.json initialization doesn't work since
one or more servers could be updating format.json on their respective
nodes. This will be addressed using distributed locking mechanism to
pick the first (lock) winner as the decider of uuids assigned to the
disks.