Improve reconnection logic, allow jitters.
Attempt a reconnect also on if disk not found.

This is needed since any network operation error
is converted to disk not found but we also need
to make sure really if disk is not available.

Additionally we also need to retry more than
once because the server might be in startup
sequence which would render other servers to
wrongly think that the server is offline.