[proposal] Support creating empty directories.
Every so often we get requirements for creating
directories/prefixes and we end up rejecting
such requirements. This PR is a propoal to show
that its possible to in-fact have empty directories
and be able to manage them like without much
backend format changes etc.

Existing APIs themselves are leveraged to provide
this behavior. Only FS backend supports this for
the time being.