require old credentials for RPC SetAuth
This commit requires the old server credentials for the SetAuth RPC
call. It will now fail if the client does not present the previous
credentials in addition to the new ones.

This commit also changes the browser to provide those credentials when
calling SetAuth (via an additional GetAuth call).

Previous to this, only an access token was necessary to set the auth.
However, access tokens are temporary while auth is permanent. Thus, the
client must have both the token and the previous auth, so as to ensure
they have access to change it.

Fixes: https://github.com/minio/minio/issues/4182