browser: add ability to delete bucket
This commit adds the ability to delete empty buckets to the brower. It
does this with a new RPC call, DeleteBucket. This requires the proper
authentication and adds tests. It only deletes a bucket that is empty =
no objects.

This also integrates it into the UI with a dropdown for each bucket.

Fixes: https://github.com/minio/minio/issues/4166