test: add Remove before Append in xl
This commit adds a removeXLMetadata call before the writeXLMetadata call.
This is necessary because the StorageAPI only implements append, and no
write (replacing prior data), and writeXLMetadata doesn't delete the old
file before appending to it. Prior to this change, there would be
multiple JSON instances in the same xl.json file, and it was a bug in
our JSON parser that it would accept the wrongly-formatted JSON.

Tests pass both before and after this change, but this change is
required to switch to a new JSON parser, which doesn't ignore this
error.