browser: move from react to vuejs
This commit moves most of the web browser from React to the VueJS
framework. This is about 90% complete. This was done because of
licensing issues with React; see:
https://github.com/minio/minio/issues/4710

This commit contains the entire rewrite. The app maintains much of its
old architecture, just this time with a new framework and a more modular
approach to the Browse component. The app state is much slimmer now, as
components should maintain as much state as they can when it isn't
relevant to other components. If it is, only then should we keep it in
our global app state.

Because this uses Vue templates instead of JSX, we can take advantage of
their superior event handling, and this is visible throughout the code
base. Additionally the separation between <template> and <script> make
the browser code much more readable and understandable.

The JSON RPC and Web action handling stays the same, as does some
actions. The UI is almost entirely unchanged; opening the new and old
browser in two tabs shows negligible difference, if you notice it.

Because this rewrite moves much of the app state from the global state,
this allows the individual components themselves to control the viewing
of the global state. For example, ObjectsList now performs object
sorting on the fly when objects are updated, as opposed to storing the
objects themselves as a sorted array. This is much better structured.
Another example is the search bucket functionality. Before, we'd have a
list of "visible buckets" and regular buckets. Now we store a search
query in the SideBar component, and buckets are dynamically searched
through when they're viewed. This should make the browser easier to
maintain in the future.

Another perk of this rewrite is that it contains the newux improvements.
This commit contains the delete bucket from the UI changes, as well as
object preview (= image thumbnails).

There are however a few things that aren't implemented yet. These
include the most important:
- object uploading
- bulk object downloads, by checked objects
- login route needs fixing, shouldn't be its own route
- dropzone and infinite scroll are not implemented
- share modal is not implemented
- unauthenticated views aren't completely ready
This is not a comprehensive list. There is a "TODO" in the code for
everything that I believe needs fixing / implementation. Besides these
TODOs, the app needs comprehensive testing before publication, but I
believe the current app state to be very close to ready.

~ with <3