Switch to non-root user for new Minio Docker deployments
Docker recommends running as non-root if the service can run without privileges
(https://docs.docker.com/engine/userguide/eng-image/dockerfile_best-practices/#user).
With this PR, we switch to non-root (minio) user while running inside a Docker
container, if there isn't an existing Minio data directory or if its an empty
data directory.

We dont switch to minio user if there is an existing data directory, this is because
we'll need to chown all the files, which may take too long, ruining the
user's experience.

Fixes: https://github.com/minio/minio/issues/4381