split hash.Reader logic into simple PUT and Gateway.
Minio has to support 3 types of etag validation/computation.

1. The client does a simple PUT with FS or XL backend.
There the server verifies that the provided content-md5 is the same
as the MD5 sum of the uploaded object. If the content-md5 is not set
the server computes it and uses it as etag.

2. The client does a multipart PUT operation with FS or XL backend.
There the server can generate a random etag (followed by a -'part-number')
per part. Currently this is done by computing the MD5 sum, too. See #4641.

3. The client does a PUT operation in gateway mode.
There the etag is computed by the cloud-provider (GCS, Azure, AWS).
In this case we should avoid computing the MD5 sum twice.

This change implements the 1. and 3. case correctly. The 2. case can be fixed
in a separate commit/PR.

Fixes #5100